refactor: moved urls to utils folder
