+++ components/home-page/bonus/bonus.tsx
@@ -6,7 +6,7 @@ import FallbackLeft from "public/assets/images/bonus/fallback-left.png";
 import Right from "public/assets/images/bonus/right.webp";
 import FallbackRight from "public/assets/images/bonus/fallback-right.png";
 import { LinkButton } from "../../link-button/link-button";
-import urls from "../../../urls/urls.json";
+import urls from "../../../utils/urls.json";
 
 import styles from "./bonus.module.css";
